Swap order of solids in G4LogicalBorderSurface
This makes the boundary behave as expected, and is most certainly due to
the fact that the photons are going from the crystal to the Teflon, and this
order matters for the behaviour of the G4LogicalBorderSurface.

However, it is encouraging that with the scintillation yield set to
100/MeV and the reflectivity set to 1 (i.e. all optical photons are
trapped until they hit the detector), the detector detects 51 photon
hits: i.e. 511 keV!